fix: header 새로고침 문제 해결
