--- src/front-end/assets/scss/main.scss
@@ -1,13 +1,0 @@
-@import "bootstrap/scss/bootstrap.scss";
-
-@font-face {
-  font-family: "DNFBitBitv2";
-  font-style: normal;
-  font-weight: 400;
-  src: url("//cdn.df.nexon.com/img/common/font/DNFBitBitv2.otf")
-    format("opentype");
-}
-
-:root {
-  --bs-body-font-family: "DNFBitBitv2";
-}
