+++ src/front-end/components/ProductList.js
@@ -1,34 +1,32 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { Container } from "react-bootstrap";
-import ProductRegister from "./ProductRegister";
+import { Link, useLocation } from "react-router-dom";
+import CreateProduct from "./CreateProduct";
 
-const ProductList = () => {
-  const [products, setProducts] = ([]); // 식품 카테고리의 데이터를 들고와야함
-  const [showProductRegister, setShowProductRegister] = useState(false); // 상품 등록/보여주기 상태관리
+const ProductList = (props) => {
+  const [products, setProducts] = useState([]); // 식품 카테고리의 데이터를 들고와야함
+  const location = useLocation();
 
-  const handleFormSubmit = (e) => {
-    e.preventDefault();
-    setShowProductRegister(true);
-  }
-
+  useEffect(() => {
+    console.log(location.search);
+    // setProducts(category);
+  })
   return (
     <Container className="border border-left-2 border-right-2 border-dark p-4 w-full ">
-      {showProductRegister ? (
-        <ProductRegister setShowProductRegister />
-      ) : (
-        <>
+      <>
+        <div className="d-flex justify-content-between">
           <h4>식품</h4>
-          <div class="card" style={{ width: "18rem" }}>
-            <img class="card-img-top" src="..." alt="Card image cap" />
-            <div class="card-body">
-              <p class="card-text">Some quick example text to build on the card title and make up the bulk of the card's content.</p>
-            </div>
-          </div>
-          <form onSubmit={handleFormSubmit}>
-            <input type="submit" value="상품등록" />
+          <form className="flex">
+            <Link to="/CreateProduct"> <input type="submit" value="상품등록" /></Link>
           </form>
-        </>
-      )}
+        </div>
+        <div className="card" style={{ width: "18rem" }}>
+          <img className="card-img-top" src="..." alt="Card image cap" />
+          <div className="card-body">
+            <p className="card-text">Some quick example text to build on the card title and make up the bulk of the card's content.</p>
+          </div>
+        </div>
+      </>
     </Container>
 
   )
