--- src/front-end/css/productList.module.css
@@ -1,3 +1,0 @@
-ul > li {
-  list-style: none;
-}
