+++ src/index.js
@@ -2,7 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-// import "./front-end/assets/scss/main.scss";
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
